+++ meinberlin/apps/geodaten/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class Config(AppConfig):
-    name = "meinberlin.apps.geodaten"
-    label = "meinberlin_geodaten"
+    name = "meinberlin.apps.geodata"
+    label = "meinberlin_geodata"