+++ meinberlin/apps/geodaten/urls.py
@@ -3,10 +3,10 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("search", views.search_addresses, name="geodaten-search-addresses"),
+    path("search", views.search_addresses, name="geodata-search-addresses"),
     path(
         "search-by-feature/<str:feature_id>/",
         views.get_address,
-        name="geodaten-get-address",
+        name="geodata-get-address",
     ),
 ]