+++ meinberlin/react/projects/ProjectsMap.jsx
@@ -68,7 +68,7 @@ const ProjectsMap = ({ items, topicChoices, ...props }) => {
         style={{ minHeight: '100%', height: '100%' }}
         className="projects-map__map"
       >
-        <SearchAndShowAddress apiUrl="/api/geodaten/search" />
+        <SearchAndShowAddress apiUrl="/api/geodata/search" />
         {/* <SearchAndShowAddress apiUrl="https://bplan-prod.liqd.net/api/addresses/" /> */}
         <ControlWrapper position="bottomleft" className="projects-map-info__wrapper">
           <ProjectsMapInfo />